refactor: loading 페이지 수정, store 이름 변경
